Create a script to copy files into a folder
This is required due to Docker can only use files that are in the same root as it is or in lower folders. Nevertheless,
to create the build of the image it is necessary to use the pyproject.toml and the uv.lock, which are located in the root
of this repository. The dockerfile is stored in lower folders, so that's the reason of the creation of the copy_files.sh