Add field validators to EmbeddingRequest model (#14)
Before the endpoint does not allow to set input parameters as None. When you try to
set the default value, you needed to not enter the parameter in the json, which was not the ideal behaviour.

Now, if you set the optional parameters as None, the default values will be taken.
If you do not write the parameter, the default value is set too.